requirements.txt
# Core web & API
fastapi==0.115.0
uvicorn==0.31.1
python-dotenv>=1.0.0

# LLM / embeddings
openai>=1.12.0
llama-index>=0.11.1
tiktoken>=0.7.0
anthropic==0.69.0

# PDF processing
pypdf>=4.2.0
PyPDF2==3.0.1
python-multipart==0.0.20



# Web scraping / HTML parsing
beautifulsoup4==4.12.3
lxml==5.2.1

# Greenlet for async / SQLAlchemy
greenlet>=3.1.0

# Optional browser automation (playwright) — only if needed
playwright>=1.47.0  # newer version compatible with greenlet>=3.1.0

